+++ cxbin/plugin/plugin3mf.h
@@ -11,7 +11,7 @@ namespace cxbin
 		virtual ~_3mfLoader();
 
 		std::string expectExtension() override;
-		bool tryLoad(FILE* f, unsigned fileSize) override;
+		bool tryLoad(FILE* f, unsigned long long fileSize) override;
 		bool load(FILE* f, unsigned fileSize, std::vector<trimesh::TriMesh*>& out, ccglobal::Tracer* tracer) override;
 	};
 }
