+++ cxbin/plugin/pluginoff.cpp
@@ -22,7 +22,7 @@ namespace cxbin
 		return "off";
 	}
 
-	bool OffLoader::tryLoad(FILE* f, unsigned fileSize)
+	bool OffLoader::tryLoad(FILE* f, unsigned long long fileSize)
 	{
 		char buf[128];
 		if (!fgets(buf, 128, f))
